Switch to using makeContent
Now we can grid.export SVG plots using the gridSVG package. However,
the plot bounds come out a little odd. I found that taking the width
height specified in the `<g id="layout.2">` and substituting the ones
in the root `<svg>` and its direct `<g>` child for it fixed the issue.
I'd like to work out what's causing the incorrect height and width tho.

Also, flags are getting trimmed when they spill over the plotting area.
I'm not sure whether this is desirable behaviour (it does not happen with
`ggsave` or when viewing directly in Rstudio).